@@ -2,7 +2,7 @@ package com.qualityunit.android.liveagentphone.service;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 
 /**
