@@ -2,7 +2,7 @@ package com.qualityunit.android.liveagentphone.store;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.text.TextUtils;
 
 import com.qualityunit.android.liveagentphone.Const;
