@@ -1,7 +1,7 @@
 package com.qualityunit.android.liveagentphone.ui.about;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
