@@ -1,8 +1,8 @@
 package com.qualityunit.android.liveagentphone.ui.common;
 
 import android.content.Context;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  * Created by rasto
