@@ -1,6 +1,6 @@
 package com.qualityunit.android.voice;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 
 import org.pjsip.pjsua2.Account;
